+++ src/runtime/tests/vm_tests.rs
@@ -1738,6 +1738,41 @@ x = num2 4, 5
 x[1]";
             test_script(script, Number(5.0.into()));
         }
+
+        #[test]
+        fn index_element_mutation() {
+            let script = "
+x = num2 4, 5
+x[1] = 99
+x[1]";
+            test_script(script, Number(99.0.into()));
+        }
+
+        #[test]
+        fn index_range_mutation() {
+            let script = "
+x = num2 4, 5
+x[0..2] = 1
+x[0] + x[1]";
+            test_script(script, Number(2.0.into()));
+        }
+
+        #[test]
+        fn index_full_range_mutation() {
+            let script = "
+x = num2 4, 5
+x[..] = 2
+x";
+            test_script(script, num2(2.0, 2.0));
+        }
+
+        #[test]
+        fn iterator_ops() {
+            let script = "
+x = num2 1, -1
+x.keep(|n| n < 0).count()";
+            test_script(script, Number(1.0.into()));
+        }
     }
 
     mod num4_test {
@@ -1818,6 +1853,41 @@ x = num4 1, -2, 3, -4
 x = num4 9, 8, 7, 6
 x[3]";
             test_script(script, Number(6.0.into()));
+        }
+
+        #[test]
+        fn index_element_mutation() {
+            let script = "
+x = num4 4, 5, 6, 7
+x[2] = 99
+x[2]";
+            test_script(script, Number(99.0.into()));
+        }
+
+        #[test]
+        fn index_range_mutation() {
+            let script = "
+x = num4 4, 5, 6, 7
+x[1..=2] = 1
+x";
+            test_script(script, num4(4.0, 1.0, 1.0, 7.0));
+        }
+
+        #[test]
+        fn index_full_range_mutation() {
+            let script = "
+x = num4 4, 5, 6, 7
+x[..] = 2
+x.sum()";
+            test_script(script, Number(8.0.into()));
+        }
+
+        #[test]
+        fn iterator_ops() {
+            let script = "
+x = num4 1, -1, 2, -2
+x.keep(|n| n > 0).count()";
+            test_script(script, Number(2.0.into()));
         }
     }